modified get_electricity_csv.py: dataframe is now adjusted so that the csv is easier to handle for the jayvee pipeline
modified get_weather_data.py: small refactor like path adjustments

added pipeline.sh: overall pipeline for the project to get a sql databse in the end.

added project_pipelineX.jv: for X = { "", 2,3,4,5}: jayvee pipeline to add tables to the sql database. For every city there is a table with temperature, and for the overall energy production in germany there is a table. #7